refactor(control): split bootstrap assembly
